Move the test helpers to their own file.
